+++ tools/build.sh
@@ -51,7 +51,7 @@ if [[ "${rustc_version}" == *"nightly"* ]] || [[ "${rustc_version}" == *"dev"* ]
         # -Z check-cfg-features requires 1.61.0-nightly
         1.3* | 1.4* | 1.5* | 1.60.*) ;;
         *)
-            check_cfg='-Z unstable-options --check-cfg=names(const_fn_trait_bound,target_feature_a)'
+            check_cfg='-Z unstable-options --check-cfg=names(atomic_maybe_uninit_const_fn_trait_bound,atomic_maybe_uninit_target_feature_lse,atomic_maybe_uninit_target_feature_a)'
             rustup ${pre_args[@]+"${pre_args[@]}"} component add clippy &>/dev/null
             subcmd=clippy
             ;;
@@ -93,6 +93,10 @@ build() {
 
     RUSTFLAGS="${RUSTFLAGS:-} ${check_cfg:-}" \
         x cargo "${args[@]}" --feature-powerset --optional-deps --workspace --ignore-private "$@"
+    if [[ "${target}" == "aarch64"* ]]; then
+        RUSTFLAGS="${RUSTFLAGS:-} ${check_cfg:-} -C target-feature=+lse" \
+            x cargo "${args[@]}" --feature-powerset --optional-deps --workspace --ignore-private --target-dir target/lse "$@"
+    fi
 }
 
 for target in "${targets[@]}"; do